+++ generate_bindings.sh
@@ -1,6 +1,7 @@
 #!/usr/bin/env bash
 
-# Always have the latest version of bindgen and rustfmt installed before using this script
+# Always have the latest version of bindgen and rustfmt installed before using this script.
+# This script require GNU sed, and does not work with the default macOS sed: `brew install gnu-sed`.
 
 set -eu
 
@@ -26,18 +27,65 @@ echo "Using macOS SDK at: $SDK_PATH"
 echo "Using $BINDGEN_VERSION"
 echo ""
 
+function cleanup_binding() {
+    local binding_path="$1"
+
+    # `Option` is in the Rust standard prelude. No need to use full path, it's just verbose.
+    sed -i 's/::core::option::Option/Option/g' "$binding_path"
+
+    # The bindings that need these types will import them directly into scope with `--raw line`
+    sed -i 's/::std::os::raw:://g' "$binding_path"
+
+    # Most low level types should not be `Copy` nor `Clone`. And `Debug` usually don't make much
+    # sense, since they are usually just pointers/binary data.
+    sed -i '/#\[derive(Debug, Copy, Clone)\]/d' "$binding_path"
+
+    # Change struct bodies to (c_void);
+    #   Search regex: {\n +_unused: \[u8; 0],\n}
+    #   Replace string: (c_void);\n
+    sed -i -e '/^pub struct .* {$/ {
+        N;N
+        s/ {\n *_unused: \[u8; 0\],\n}/(c_void);\n/
+    }' "$binding_path"
+
+    # Remove all }\nextern "C" { to condense code a bit
+    #   Search regex: }\nextern "C" {
+    #   Replace string:
+    sed -i -e '/^extern "C" {$/ {
+        :loop
+        n
+        /^}$/! b loop
+        /^}$/ {
+            N
+            t reset_condition_flags
+            :reset_condition_flags
+            s/}\nextern "C" {//
+            t loop
+        }
+    }' "$binding_path"
+
+    rustfmt +nightly "$binding_path"
+}
+
+BINDGEN_COMMON_ARGUMENTS=(
+    --no-doc-comments
+    --use-core
+    --no-layout-tests
+    --raw-line "// Generated using:"
+    --raw-line "// $BINDGEN_VERSION"
+    --raw-line "// macOS SDK $SDK_VERSION."
+    --raw-line ""
+)
+
 echo "Generating bindings for $PREFERENCES_HEADER_PATH"
 bindgen \
-    --no-doc-comments \
+    "${BINDGEN_COMMON_ARGUMENTS[@]}" \
     --whitelist-function "SCPreferences.*" \
     --blacklist-type "(__)?CF.*" \
     --blacklist-type "Boolean" \
     --blacklist-type "dispatch_queue_[ts]" \
     --blacklist-type "(AuthorizationOpaqueRef|__SCPreferences)" \
-    --raw-line "// Generated using:" \
-    --raw-line "// $BINDGEN_VERSION" \
-    --raw-line "// macOS SDK $SDK_VERSION." \
-    --raw-line "" \
+    --raw-line "use core::ffi::c_void;" \
     --raw-line "use core_foundation_sys::array::CFArrayRef;" \
     --raw-line "use core_foundation_sys::base::{Boolean, CFIndex, CFAllocatorRef, CFTypeID};" \
     --raw-line "use core_foundation_sys::data::CFDataRef;" \
@@ -46,7 +94,6 @@ bindgen \
     --raw-line "use core_foundation_sys::runloop::CFRunLoopRef;" \
     --raw-line "" \
     --raw-line "use dispatch_queue_t;" \
-    --raw-line "use libc::c_void;" \
     --raw-line "" \
     --raw-line "pub type AuthorizationOpaqueRef = c_void;" \
     --raw-line "pub type __SCPreferences = c_void;" \
@@ -55,7 +102,7 @@ bindgen \
     -I$SDK_PATH/usr/include \
     -F$FRAMEWORK_PATH
 
-rustfmt $PREFERENCES_BINDING_PATH
+cleanup_binding $PREFERENCES_BINDING_PATH
 
 echo ""
 echo ""
@@ -63,16 +110,13 @@ echo "Generating bindings for $DYNAMIC_STORE_HEADER_PATH"
 sleep 2
 
 bindgen \
-    --no-doc-comments \
+    "${BINDGEN_COMMON_ARGUMENTS[@]}" \
     --whitelist-function "SCDynamicStore.*" \
     --whitelist-var "kSCDynamicStore.*" \
     --blacklist-type "(__)?CF.*" \
     --blacklist-type "Boolean" \
     --blacklist-type "dispatch_queue_[ts]" \
-    --raw-line "// Generated using:" \
-    --raw-line "// $BINDGEN_VERSION" \
-    --raw-line "// macOS SDK $SDK_VERSION." \
-    --raw-line "" \
+    --raw-line "use core::ffi::c_void;" \
     --raw-line "use core_foundation_sys::array::CFArrayRef;" \
     --raw-line "use core_foundation_sys::base::{Boolean, CFIndex, CFAllocatorRef, CFTypeID};" \
     --raw-line "use core_foundation_sys::string::CFStringRef;" \
@@ -86,7 +130,7 @@ bindgen \
     -I$SDK_PATH/usr/include \
     -F$FRAMEWORK_PATH
 
-rustfmt $DYNAMIC_STORE_BINDING_PATH
+cleanup_binding $DYNAMIC_STORE_BINDING_PATH
 
 echo ""
 echo ""
@@ -94,7 +138,7 @@ echo "Generating bindings for $NETWORK_CONFIGURATION_HEADER_PATH"
 sleep 2
 
 bindgen \
-    --no-doc-comments \
+    "${BINDGEN_COMMON_ARGUMENTS[@]}" \
     --whitelist-function "SCNetwork.*" \
     --whitelist-function "SCBondInterface.*" \
     --whitelist-var "kSC(NetworkInterface|BondStatus).*" \
@@ -104,10 +148,7 @@ bindgen \
     --blacklist-type "Boolean" \
     --blacklist-type "(sockaddr|socklen_t|sa_family_t|__darwin_socklen_t|__uint.*_t)" \
     --blacklist-type "(__)?SCPreferences.*" \
-    --raw-line "// Generated using:" \
-    --raw-line "// $BINDGEN_VERSION" \
-    --raw-line "// macOS SDK $SDK_VERSION." \
-    --raw-line "" \
+    --raw-line "use core::ffi::c_void;" \
     --raw-line "use core_foundation_sys::array::CFArrayRef;" \
     --raw-line "use core_foundation_sys::base::{Boolean, CFIndex, CFAllocatorRef, CFTypeID};" \
     --raw-line "use core_foundation_sys::string::CFStringRef;" \
@@ -115,7 +156,7 @@ bindgen \
     --raw-line "use core_foundation_sys::runloop::CFRunLoopRef;" \
     --raw-line "" \
     --raw-line "use dispatch_queue_t;" \
-    --raw-line "use libc::{c_void, c_char, c_int, sockaddr};" \
+    --raw-line "use libc::{c_char, c_int, sockaddr, socklen_t};" \
     --raw-line "use preferences::SCPreferencesRef;" \
     --raw-line "" \
     --raw-line "pub type __SCNetworkReachability = c_void;" \
@@ -130,7 +171,7 @@ bindgen \
     -I$SDK_PATH/usr/include \
     -F$FRAMEWORK_PATH
 
-rustfmt $NETWORK_CONFIGURATION_BINDING_PATH
+cleanup_binding $NETWORK_CONFIGURATION_BINDING_PATH
 
 echo ""
 echo ""
@@ -138,13 +179,9 @@ echo "Generating bindings for $SCHEMA_DEFINITIONS_HEADER_PATH"
 sleep 2
 
 bindgen \
-    --no-doc-comments \
+    "${BINDGEN_COMMON_ARGUMENTS[@]}" \
     --whitelist-var "kSC.*" \
     --blacklist-type "(__)?CF.*" \
-    --raw-line "// Generated using:" \
-    --raw-line "// $BINDGEN_VERSION" \
-    --raw-line "// macOS SDK $SDK_VERSION." \
-    --raw-line "" \
     --raw-line "use core_foundation_sys::string::CFStringRef;" \
     --raw-line "" \
     -o $SCHEMA_DEFINITIONS_BINDING_PATH \
@@ -152,4 +189,4 @@ bindgen \
     -I$SDK_PATH/usr/include \
     -F$FRAMEWORK_PATH
 
-rustfmt $SCHEMA_DEFINITIONS_BINDING_PATH
+cleanup_binding $SCHEMA_DEFINITIONS_BINDING_PATH